Operational predictions for rbf dkl cox proc

+++ pgf_kernel_experiments/runners/exact_multi_gp_runner.py
@@ -47,11 +47,11 @@ class ExactMultiGPRunner:
     def predict(self, test_x):
         return [self.single_runners[i].predict(test_x) for i in range(self.num_gps())]
 
-    def assess(self, predictions, test_y, metrics, verbose=True):
-        scores = torch.empty([self.num_gps(), len(metrics)], dtype=test_y.dtype, device=test_y.device)
+    def assess(self, predictions, test_y, metrics, dtype=torch.float64, verbose=True):
+        scores = torch.empty([self.num_gps(), len(metrics)], dtype=dtype, device=test_y.device)
 
         for i in range(self.num_gps()):
-            scores[i, :] = self.single_runners[i].assess(predictions[i], test_y, metrics, verbose=verbose)
+            scores[i, :] = self.single_runners[i].assess(predictions[i], test_y, metrics, dtype=dtype, verbose=verbose)
 
         return scores
