+++ pgf_kernel_experiments/runners/exact_single_dkl_runner.py
@@ -64,8 +64,8 @@ class ExactSingleDKLRunner:
 
         return predictions
 
-    def assess(self, predictions, test_y, metrics, verbose=True):
-        scores = torch.empty([len(metrics)], dtype=test_y.dtype, device=test_y.device)
+    def assess(self, predictions, test_y, metrics, dtype=torch.float64, verbose=True):
+        scores = torch.empty([len(metrics)], dtype=dtype, device=test_y.device)
 
         if verbose:
             msg = ', '.join(['{:.6f}']*len(metrics))
